+++ src/main/java/insane96mcp/insanesurvivaloverhaul/data/generator/ISOBlockTagsProvider.java
@@ -27,11 +27,11 @@ public class ISOBlockTagsProvider extends BlockTagsProvider {
     //public static final TagKey<Block> ASPEN_LEAVES = create("aspen_leaves");
     //public static final TagKey<Block> PLUM_LEAVES = create("plum_leaves");
 
-    //public static final TagKey<Block> HUD_CARDINAL_DIRECTION = create("hud/cardinal_direction");
-    //public static final TagKey<Block> HUD_SEASON = create("hud/season");
-    //public static final TagKey<Block> HUD_DEPTH = create("hud/depth");
-    //public static final TagKey<Block> HUD_TIME = create("hud/time");
-    //public static final TagKey<Block> HUD_BIOME = create("hud/biome");
+    public static final TagKey<Block> HUD_CARDINAL_DIRECTION = create("hud/cardinal_direction");
+    public static final TagKey<Block> HUD_SEASON = create("hud/season");
+    public static final TagKey<Block> HUD_DEPTH = create("hud/depth");
+    public static final TagKey<Block> HUD_TIME = create("hud/time");
+    public static final TagKey<Block> HUD_BIOME = create("hud/biome");
 
     public ISOBlockTagsProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, String modId, @Nullable ExistingFileHelper existingFileHelper){
         super(output, lookupProvider, modId, existingFileHelper);
@@ -39,7 +39,8 @@ public class ISOBlockTagsProvider extends BlockTagsProvider {
 
     @Override
     protected void addTags(HolderLookup.Provider provider) {
-
+        //tag(HUD_TIME)
+                //.addOptional(ForgeRegistries.BLOCKS.getKey(ModRegistry.CLOCK_BLOCK.get()));
     }
 
     public static TagKey<Block> create(String tagName) {